fix: refactored routes file
100% codecov on routes file and renamed it to fit better what it is.
